requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
django-dotenv==1.4.2
idna==3.3
Pillow==9.1.0
psycopg2-pgevents==0.1.1
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
twilio==7.9.0
tzdata==2022.1
urllib3==1.26.9
